requirements.txt
PyID3>=0.5.5
PIL>=1.1.7
